requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
openai>=1.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
